added js to make session options for review form
Adds session options according to the current month and year. This
will have some difficulty handling future exceptions. Currently assume
each first session of the year starts in April and ends in September,
and second sessions of the year start in October and end in March.

Does not add new option until 1 month after session starts. This is
a result of a simple check (if current month > session start month)
but could be tweaked. If we change this we need to be mindful of the
month wrapping back to 0.

Uses jQuery (which is loaded for bootstrap) which requires that we
wait for it to load first. This is why the SessionOptions object is
instantiated in the window.onload function. I did this because jQuery
is nice for selecting and adding elements. This causes no performance
issues so far as I know. Tested with 'lighthouse' on chrome and it
received a 97 performance score.